Adiciona processo completo de CRUD com os valores de aluguel e moto

+++ Application/Orchestrators/Motorcycles/MotorcycleDeleteOrchestrator.cs
@@ -1,6 +1,7 @@
 ﻿using MotosAluguel.Application.Commons;
 using MotosAluguel.Application.Interfaces.Orchestrators.Motorcycles;
 using MotosAluguel.Domain.Interfaces.Repositories.Motorcyles;
+using MotosAluguel.Domain.Interfaces.Validators.Motorcycles;
 
 namespace MotosAluguel.Application.Orchestrators.Motorcycles;
 
@@ -8,15 +9,28 @@ public class MotorcycleDeleteOrchestrator : IMotorcycleDeleteOrchestrator
 {
     private readonly IMotorcycleWriterRepository _repository;
 
-    public MotorcycleDeleteOrchestrator(IMotorcycleWriterRepository repository)
+    private readonly IMotorcycleDeleteValidator _validator;
+
+    public MotorcycleDeleteOrchestrator(
+        IMotorcycleWriterRepository repository,
+        IMotorcycleDeleteValidator motorcycleDeleteValidator)
     {
         _repository = repository;
+        _validator = motorcycleDeleteValidator;
     }
 
     public async Task<OperationResult<bool>> RunAsync(string id)
     {
-        await _repository.DeleteAsync(id);
+        var isValid = await _validator.ValidateAsync(id);
 
-        return OperationResult<bool>.Ok(true);
+        if (isValid)
+        {
+            var resultValid = await _repository.DeleteAsync(id);
+
+            if(resultValid)
+                return OperationResult<bool>.Ok();
+        }
+
+        return OperationResult<bool>.Fail("Dados inválidos");
     }
 }
